[Travis] Fix #30 Refactorinng functions in 2 Entities files
Use Inheritance Mapping from Doctrine with @MappedSuperclass annotation
