+++ src/Glsr/GestockBundle/Entity/Company.php
@@ -16,8 +16,6 @@
 namespace Glsr\GestockBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
-use Misd\PhoneNumberBundle\Validator\Constraints\PhoneNumber as AssertPhoneNumber;
-use Symfony\Component\Validator\Constraints as Assert;
 
 /**
  * Company Entité Company.
@@ -27,24 +25,8 @@ use Symfony\Component\Validator\Constraints as Assert;
  * @ORM\Table(name="gs_company")
  * @ORM\Entity(repositoryClass="Glsr\GestockBundle\Entity\CompanyRepository")
  */
-class Company
+class Company extends Contact
 {
-    /**
-     * @var int Id de l'entreprise
-     *
-     * @ORM\Column(name="id", type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    private $idComp;
-
-    /**
-     * @var string Nom de l'entreprise
-     *
-     * @ORM\Column(name="name", type="string", length=255)
-     */
-    private $name;
-
     /**
      * @var string Statut de l'entreprise
      *
@@ -52,106 +34,6 @@ class Company
      */
     private $status;
 
-    /**
-     * @var string Adresse de l'entreprise
-     *
-     * @ORM\Column(name="adress", type="string", length=255)
-     */
-    private $adress;
-
-    /**
-     * @var string Code postal
-     *
-     * @ORM\Column(name="zipcode", type="string", length=11)
-     */
-    private $zipcode;
-
-    /**
-     * @var string Ville
-     *
-     * @ORM\Column(name="town", type="string", length=255)
-     */
-    private $town;
-
-    /**
-     * @var phone_number Téléphone de l'entreprise
-     *
-     * @ORM\Column(name="phone", type="phone_number")
-     * @Assert\NotBlank()
-     * @AssertPhoneNumber(defaultRegion="FR")
-     */
-    private $phone;
-
-    /**
-     * @var phone_number Fax de l'entreprise
-     *
-     * @ORM\Column(name="fax", type="phone_number")
-     * @Assert\NotBlank()
-     * @AssertPhoneNumber(defaultRegion="FR")
-     */
-    private $fax;
-
-    /**
-     * @var string email de l'entreprise
-     *
-     * @ORM\Column(name="mail", type="string", length=255)
-     * @Assert\NotBlank()
-     * @Assert\Email(
-     *     message = "'{{ value }}' n'est pas un email valide.",
-     *     checkMX = true
-     * )
-     */
-    private $mail;
-
-    /**
-     * @var string Contact de votre entreprise
-     *
-     * @ORM\Column(name="contact", type="string", length=255)
-     */
-    private $contact;
-
-    /**
-     * @var phone_number Gsm de votre entreprise
-     *
-     * @ORM\Column(name="gsm", type="phone_number")
-     * @Assert\NotBlank()
-     * @AssertPhoneNumber(defaultRegion="FR")
-     */
-    private $gsm;
-
-    /**
-     * Get id.
-     *
-     * @return int
-     */
-    public function getId()
-    {
-        return $this->idComp;
-    }
-
-    /**
-     * Set name.
-     *
-     * @param string $name Nom de l'entreprise
-     *
-     * @return Company
-     */
-    public function setName($name)
-    {
-        $this->name = $name;
-
-        return $this;
-    }
-
-    /**
-     * Get name.
-     *
-     * @return string
-     */
-    public function getName()
-    {
-        return $this->name;
-    }
 
     /**
      * Set status.
@@ -175,197 +57,5 @@ class Company
     public function getStatus()
     {
         return $this->status;
-    }
-
-    /**
-     * Set adress.
-     *
-     * @param string $adress Adresse
-     *
-     * @return Company
-     */
-    public function setAdress($adress)
-    {
-        $this->adress = $adress;
-
-        return $this;
-    }
-
-    /**
-     * Get adress.
-     *
-     * @return string
-     */
-    public function getAdress()
-    {
-        return $this->adress;
-    }
-
-    /**
-     * Set zipcode.
-     *
-     * @param string $zipcode Code postal
-     *
-     * @return Company
-     */
-    public function setZipcode($zipcode)
-    {
-        $this->zipcode = $zipcode;
-
-        return $this;
-    }
-
-    /**
-     * Get zipcode.
-     *
-     * @return string
-     */
-    public function getZipcode()
-    {
-        return $this->zipcode;
-    }
-
-    /**
-     * Set town.
-     *
-     * @param string $town Ville
-     *
-     * @return Company
-     */
-    public function setTown($town)
-    {
-        $this->town = $town;
-
-        return $this;
-    }
-
-    /**
-     * Get town.
-     *
-     * @return string
-     */
-    public function getTown()
-    {
-        return $this->town;
-    }
-
-    /**
-     * Set phone.
-     *
-     * @param string $phone Téléphone
-     *
-     * @return Company
-     */
-    public function setPhone($phone)
-    {
-        $this->phone = $phone;
-
-        return $this;
-    }
-
-    /**
-     * Get phone.
-     *
-     * @return string
-     */
-    public function getPhone()
-    {
-        return $this->phone;
-    }
-
-    /**
-     * Set fax.
-     *
-     * @param string $fax Fax
-     *
-     * @return Company
-     */
-    public function setFax($fax)
-    {
-        $this->fax = $fax;
-
-        return $this;
-    }
-
-    /**
-     * Get fax.
-     *
-     * @return string
-     */
-    public function getFax()
-    {
-        return $this->fax;
-    }
-
-    /**
-     * Set mail.
-     *
-     * @param string $mail Adresse Email
-     *
-     * @return Company
-     */
-    public function setMail($mail)
-    {
-        $this->mail = $mail;
-
-        return $this;
-    }
-
-    /**
-     * Get mail.
-     *
-     * @return string
-     */
-    public function getMail()
-    {
-        return $this->mail;
-    }
-
-    /**
-     * Set contact.
-     *
-     * @param string $contact Nom du contact
-     *
-     * @return Company
-     */
-    public function setContact($contact)
-    {
-        $this->contact = $contact;
-
-        return $this;
-    }
-
-    /**
-     * Get contact.
-     *
-     * @return string
-     */
-    public function getContact()
-    {
-        return $this->contact;
-    }
-
-    /**
-     * Set gsm.
-     *
-     * @param string $gsm Numéro de téléphone du contact
-     *
-     * @return Company
-     */
-    public function setGsm($gsm)
-    {
-        $this->gsm = $gsm;
-
-        return $this;
-    }
-
-    /**
-     * Get gsm.
-     *
-     * @return string
-     */
-    public function getGsm()
-    {
-        return $this->gsm;
     }
 }
